voice-bridge/requirements.txt
discord.py[voice]>=2.3.0
elevenlabs>=1.0.0
PyNaCl>=1.5.0
python-dotenv>=1.0.0
